[LogicCombine 2/?] Add weight into node then use it to determine if we can combine the case need to create new value.
In this patch we involve two member variables into logical node: `Weight` , `OneUseWeight`.
`Weight` maintain how many instructions can we get the value based on original IR code.
`OneUseWeight`means if we rebuild the final instruction, how many instructions we need to keep for the child node without one-use.
After that we can measure the cost if we rebuild the instruction.

Differential Revision: https://reviews.llvm.org/D143046